Work on the daemon. PROMPT_COMMAND now working.
Note that sinking to files is a temporary measure, because once all
the data is being recorded correctly we need to sessionise it.

Trap command is still not working, need to explore why. One annoying
thing is that it fires on the prompt command events, which needs to
be sorted.