Implement redis destination.
You can now collect data on a redis server using the protocol
redis://host:port. At the moment, there is no option to specify a
password, but this would be easy to add (just a parser change.)

It's not clear what the correct behaviour should be when the server
disconnects. I think the best we could do it just to print out
a warning to the terminal if there is no server, or if it disconnects.
Should probably be able to reconnect if the server blips.